+++ commands/suggs/denied-sugg.js
@@ -1,4 +1,4 @@
-const Discord = require('discord.js');
+const { Discord, PermissionsBitField } = require('discord.js');
 const connection = require('../../database.js');
 const config = require('../../config/config.json');
 
@@ -90,6 +90,20 @@ module.exports = {
                     console.log(error);
                 }
             const chnnel = await message.guild.channels.cache.find(c => c.name === 'suggestions');
+            const botPermissionsIn = message.guild.members.me.permissionsIn(chnnel);
+            if(!botPermissionsIn.has(PermissionsBitField.Flags.SendMessages)) return message.author.send(`I can\'t send messages in that channel. I need to have the \`SEND MESSAGES\` permission for that channel. A mod or guild owner will need to update this. If you are seeing this in error, please run the \`${prefix}report\` command.`);
+
+            const botPerms = [PermissionsBitField.Flags.SendMessages, PermissionsBitField.Flags.ViewChannel, PermissionsBitField.Flags.ReadMessageHistory, ]
+            let v = 0;
+            for(const i of botPerms) {
+                if(!message.guild.members.me.permissionsIn(chnnel).has(i)) {
+                    v++
+                }
+                if(v == botPerms.length) {
+                    message.react('❌');
+                    return message.author.send('I do not have the necessary permissions for this channel. I need \`Read Message History, View Channel, and Send Messages.\`');
+                }
+            }
             chnnel.messages.fetch(msgId).then(message => {
                     message.delete();
                 }
